+++ changelog.md
@@ -38,6 +38,7 @@
   - Mapper lifetimes (transient, scoped, singleton)
 - #658 Limit number of product clones to generate either via UI or code (or both)
 - #736 Backend > Product grid: Search By SKU should also search for GTIN & MPN
+- #682 Allow to set the default email account per shop.
 
 ### Bugfixes